DRTII-1050 Start adding tests around setUpdatedAt
