+++ server/src/main/scala/uk/gov/homeoffice/drt/testsystem/db/AggregateDbH2.scala
@@ -20,6 +20,7 @@ object AggregateDbH2 extends AggregatedDbTables {
     arrivalStats,
     processedZip,
     processedJson,
+    statusDaily,
     voyageManifestPassengerInfo,
   )
